requirements.txt
qiskit[visualization]~=2.3.0
qiskit-aer-gpu-cu11~=0.17.0
qiskit-ibm-runtime~=0.41.1
tqdm~=4.67.1
flake8~=7.3.0
pytest~=8.4.1
qutip~=5.2.0
scikit-learn~=1.7.1
numpy~=2.2.6
matplotlib~=3.10.3
PyYAML>=6.0
omegaconf~=2.3.0
scipy~=1.16.0
yaml~=0.2.5
hydra-core~=1.3.2